Changed options value 'hats' for grammar purposes
Easier than to re-write the whole JS logic for the 'other' category when
the guess is made and prompt displayed for the user, as we're using booleans
rather than strings for that category, at least for my skill level.
Though 'facehair' isn't really proper English.
Compatible use if in Swedish, however, so I'll just let that be for now... :)